requirements.txt
fastapi
uvicorn
gunicorn          # ✅ for production deployment
jinja2
pytest
requests
pre-commit
python-dotenv
aiofiles
python-multipart
pillow
pytesseract
pydantic
opencv-python     # 🚨 REQUIRED for preprocessing
streamlit       # optional — uncomment if using
httpx           # optional — better async client
filetype        # optional — extra file validation